fix：update method call location of EventTrack.Transaction
